+++ test/devices_tests/expose_sensor_test.py
@@ -221,7 +221,7 @@ class TestExposeSensor:
     async def test_cooldown(self, time_travel):
         """Test cooldown."""
         xknx = XKNX()
-        xknx.knxip_interface = AsyncMock()
+        xknx.cemi_handler = AsyncMock()
         await xknx.telegram_queue.start()
 
         expose_sensor_cd = ExposeSensor(
@@ -241,7 +241,7 @@ class TestExposeSensor:
         await expose_sensor_cd.set(21.0)
         await expose_sensor_no_cd.set(21.0)
         await time_travel(0)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -257,13 +257,13 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
 
         # don't send telegram with same payload twice if cooldown is active
         await expose_sensor_cd.set(21.0)
         await expose_sensor_no_cd.set(21.0)
         await time_travel(0)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -273,17 +273,17 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
 
         await time_travel(10)
         assert xknx.telegrams.qsize() == 0
-        xknx.knxip_interface.send_telegram.assert_not_called()
+        xknx.cemi_handler.send_telegram.assert_not_called()
 
         # different payload after cooldown
         await expose_sensor_cd.set(10.0)
         await expose_sensor_no_cd.set(10.0)
         await time_travel(0)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -299,12 +299,12 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
         # different payload immediately (payload of 3.111 equals 3.11)
         await expose_sensor_cd.set(3.111)
         await expose_sensor_no_cd.set(3.111)
         await time_travel(0)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -314,11 +314,11 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
         assert expose_sensor_cd._cooldown_latest_value == 3.111
         assert expose_sensor_cd.sensor_value.value == 10
         await time_travel(10)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -328,17 +328,17 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
         assert expose_sensor_cd._cooldown_latest_value == 3.11
         assert expose_sensor_cd.sensor_value.value == 3.11
         await time_travel(10)
-        xknx.knxip_interface.send_telegram.assert_not_called()
+        xknx.cemi_handler.send_telegram.assert_not_called()
 
         # reading unsent value
         await expose_sensor_cd.set(10)
         #   first send new value
         await time_travel(0)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -348,14 +348,14 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
         #   in cooldown with a new value - receiving a read request
         await expose_sensor_cd.set(21)
         await expose_sensor_cd.process(
             Telegram(GroupAddress("1/2/3"), payload=GroupValueRead())
         )
         await time_travel(0)
-        xknx.knxip_interface.send_telegram.assert_has_calls(
+        xknx.cemi_handler.send_telegram.assert_has_calls(
             [
                 call(
                     Telegram(
@@ -365,8 +365,8 @@ class TestExposeSensor:
                 ),
             ]
         )
-        xknx.knxip_interface.send_telegram.reset_mock()
+        xknx.cemi_handler.send_telegram.reset_mock()
         #   after cooldown - new value not sent again (already in GroupValueResponse)
         await time_travel(10)
-        xknx.knxip_interface.send_telegram.assert_not_called()
+        xknx.cemi_handler.send_telegram.assert_not_called()
         await xknx.stop()