+++ xknx/cemi/__init__.py
@@ -1,0 +1,3 @@
+"""Module for handling CEMI Frames."""
+# flake8: noqa
+from .cemi_handler import CEMIHandler