+++ xknx/io/routing.py
@@ -24,10 +24,10 @@ from xknx.knxip import (
     RoutingIndication,
     RoutingLostMessage,
 )
-from xknx.telegram import IndividualAddress, Telegram, TelegramDirection
+from xknx.telegram import IndividualAddress
 
 from .const import DEFAULT_INDIVIDUAL_ADDRESS, DEFAULT_MCAST_GRP, DEFAULT_MCAST_PORT
-from .interface import Interface, TelegramCallbackType
+from .interface import CEMICallbackType, Interface
 from .ip_secure import SecureGroup
 from .transport import KNXIPTransport, UDPTransport
 
@@ -140,7 +140,7 @@ class Routing(Interface):
         self,
         xknx: XKNX,
         individual_address: IndividualAddress | None,
-        telegram_received_callback: TelegramCallbackType,
+        cemi_received_callback: CEMICallbackType,
         local_ip: str,
         multicast_group: str = DEFAULT_MCAST_GRP,
         multicast_port: int = DEFAULT_MCAST_PORT,
@@ -148,7 +148,7 @@ class Routing(Interface):
         """Initialize Routing class."""
         self.xknx = xknx
         self.individual_address = individual_address or DEFAULT_INDIVIDUAL_ADDRESS
-        self.telegram_received_callback = telegram_received_callback
+        self.cemi_received_callback = cemi_received_callback
         self.local_ip = local_ip
         self.multicast_group = multicast_group
         self.multicast_port = multicast_port
@@ -217,18 +217,17 @@ class Routing(Interface):
     #
     ##################
 
-    async def send_telegram(self, telegram: Telegram) -> None:
-        """Send Telegram to routing connected device."""
-        cemi = CEMIFrame.init_from_telegram(
-            telegram=telegram,
-            code=CEMIMessageCode.L_DATA_IND,
-            src_addr=self.individual_address,
-        )
-        cemi_logger.debug("Outgoing CEMI: %s", cemi)
+    async def send_cemi(self, cemi: CEMIFrame) -> None:
+        """Send CEMIFrame to the network."""
+        # send L_DATA_IND to network, create L_DATA_CON locally for routing
+        cemi.code = CEMIMessageCode.L_DATA_IND
         routing_indication = RoutingIndication(raw_cemi=cemi.to_knx())
 
         async with self._flow_control.throttle():
             self._send_knxipframe(KNXIPFrame.init_from_body(routing_indication))
+
+        cemi.code = CEMIMessageCode.L_DATA_CON
+        self.cemi_received_callback(cemi)
 
     def _send_knxipframe(self, knxipframe: KNXIPFrame) -> None:
         """Send KNXIPFrame to connected routing device."""
@@ -262,16 +261,13 @@ class Routing(Interface):
         cemi = CEMIFrame()
         try:
             cemi.from_knx(routing_indication.raw_cemi)
-            cemi_logger.debug("Incoming CEMI: %s", cemi)
         except UnsupportedCEMIMessage as unsupported_cemi_err:
             logger.warning("CEMI not supported: %s", unsupported_cemi_err)
             return
         if cemi.src_addr == self.individual_address:
-            logger.debug("Ignoring own packet")
+            logger.debug("Ignoring own packet %s", cemi)
             return
-        telegram = cemi.telegram
-        telegram.direction = TelegramDirection.INCOMING
-        self.telegram_received_callback(telegram)
+        self.cemi_received_callback(cemi)
 
 
 class SecureRouting(Routing):
@@ -283,7 +279,7 @@ class SecureRouting(Routing):
         self,
         xknx: XKNX,
         individual_address: IndividualAddress | None,
-        telegram_received_callback: TelegramCallbackType,
+        cemi_received_callback: CEMICallbackType,
         local_ip: str,
         backbone_key: bytes,
         latency_ms: int | None = None,
@@ -296,7 +292,7 @@ class SecureRouting(Routing):
         super().__init__(
             xknx,
             individual_address=individual_address,
-            telegram_received_callback=telegram_received_callback,
+            cemi_received_callback=cemi_received_callback,
             local_ip=local_ip,
             multicast_group=multicast_group,
             multicast_port=multicast_port,